+++ src/main/java/prismatic/shards/stellarity/datagen/loot_table/FishingLootTableProvider.java
@@ -9,9 +9,11 @@ import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.tags.EnchantmentTags;
 import net.minecraft.world.item.component.FireworkExplosion;
+import net.minecraft.world.level.biome.Biomes;
 import net.minecraft.world.level.storage.loot.LootTable;
 import net.minecraft.world.level.storage.loot.parameters.LootContextParamSets;
 import org.jspecify.annotations.NonNull;
+import prismatic.shards.stellarity.key.StellarityBiomes;
 import prismatic.shards.stellarity.util.LootUtil;
 
 import java.util.HashMap;
@@ -31,21 +33,34 @@ public class FishingLootTableProvider extends SimpleFabricLootTableSubProvider {
 	public FishingLootTableProvider(FabricPackOutput output, CompletableFuture<HolderLookup.Provider> registryLookup) {
 		super(output, registryLookup, LootContextParamSets.FISHING);
 		this.registryLookup = registryLookup;
-
 	}
 
-
-	public static final HashMap<ResourceKey<LootTable>, LootTable.Builder> LOOT_TABLES = new HashMap<>();
-
-
-	public static void define(HolderLookup.Provider lookup) {
-		LOOT_TABLES.put(VOID_FISHING_FISH, lootTable().withPool(pool()
+	@Override
+	public void generate(@NonNull BiConsumer<ResourceKey<LootTable>, LootTable.Builder> consumer) {
+		var lookup = registryLookup.join();
+		var biomes = lookup.lookupOrThrow(Registries.BIOME);
+		consumer.accept(VOID_FISHING_FISH, lootTable().withPool(pool()
 			.add(item(ENDER_KOI).setWeight(15).apply(count(range(2, 4))))
-			.add(item(CRYSTAL_HEARTFISH).setWeight(4)
-			)));
+			.add(item(CRYSTAL_HEARTFISH).setWeight(4))
+			.add(tableLoot(VOID_FISHING_LOCATION_THE_HALLOW).setWeight(30).when(biome(biomes.getOrThrow(StellarityBiomes.THE_HALLOW))))
+			.add(tableLoot(VOID_FISHING_LOCATION_FLESH_TUNDRA).setWeight(30).when(biome(biomes.getOrThrow(StellarityBiomes.FLESH_TUNDRA))))
+			.add(tableLoot(VOID_FISHING_LOCATION_WARPED_MARSH).setWeight(30).when(biome(biomes.getOrThrow(StellarityBiomes.WARPED_MARSH))))
+			.add(tableLoot(VOID_FISHING_LOCATION_FROZEN_SPIKES).setWeight(30).when(biome(biomes.getOrThrow(StellarityBiomes.FROZEN_SPIKES))))
+			.add(tableLoot(VOID_FISHING_LOCATION_VANILLA_BIOMES).when(any(
+				biome(biomes.getOrThrow(Biomes.END_BARRENS)),
+				biome(biomes.getOrThrow(Biomes.END_HIGHLANDS)),
+				biome(biomes.getOrThrow(Biomes.END_MIDLANDS)),
+				biome(biomes.getOrThrow(Biomes.SMALL_END_ISLANDS))
+			)))
+			.add(tableLoot(VOID_FISHING_LOCATION_FIERY_HILLS).setWeight(30).when(biome(biomes.getOrThrow(StellarityBiomes.FIERY_HILLS))))
+			.add(tableLoot(VOID_FISHING_LOCATION_AMETHYST_BIOMES).when(any(
+				biome(biomes.getOrThrow(StellarityBiomes.AMETHYST_FOREST)),
+				biome(biomes.getOrThrow(StellarityBiomes.CRYSTAL_CRAGS))
+			)))
+			.add(tableLoot(VOID_FISHING_LOCATION_PRISMARINE_FOREST).setWeight(30).when(biome(biomes.getOrThrow(StellarityBiomes.PRISMARINE_FOREST))))
+		));
 
-
-		LOOT_TABLES.put(VOID_FISHING_JUNK, lootTable().withPool(pool()
+		consumer.accept(VOID_FISHING_JUNK, lootTable().withPool(pool()
 			.add(item(CRYING_OBSIDIAN).apply(count(range(1, 3))))
 			.add(item(OBSIDIAN).apply(count(range(1, 4))))
 			.add(item(END_STONE).apply(count(range(3, 9))))
@@ -73,7 +88,7 @@ public class FishingLootTableProvider extends SimpleFabricLootTableSubProvider {
 			.add(item(BLACK_DYE))
 		));
 
-		LOOT_TABLES.put(VOID_FISHING_TREASURE, lootTable().withPool(pool()
+		consumer.accept(VOID_FISHING_TREASURE, lootTable().withPool(pool()
 			.add(item(EYE_ARMOR_TRIM_SMITHING_TEMPLATE).setWeight(7).apply(count(range(1, 2))))
 			.add(item(SPIRE_ARMOR_TRIM_SMITHING_TEMPLATE).setWeight(7).apply(count(range(1, 2))))
 			.add(item(END_CRYSTAL).setWeight(5).apply(count(range(1, 2))))
@@ -86,24 +101,81 @@ public class FishingLootTableProvider extends SimpleFabricLootTableSubProvider {
 			))
 		));
 
-		LOOT_TABLES.put(VOID_FISHING_FISHER_OF_VOIDS, lootTable().withPool(pool().add(item(FISHER_OF_VOIDS)
+		consumer.accept(VOID_FISHING_FISHER_OF_VOIDS, lootTable().withPool(pool().add(item(FISHER_OF_VOIDS)
 			.apply(damage(0.15f, 0.75f))
 		)));
 
-		LOOT_TABLES.put(VOID_FISHING_EVENT, lootTable().withPool(pool()
+
+		consumer.accept(VOID_FISHING_EVENT, lootTable().withPool(pool()
 			.add(lootTable(VOID_FISHING_JUNK).setWeight(15).setQuality(-2))
 			.add(lootTable(VOID_FISHING_TREASURE).setWeight(4).setQuality(2))
 			.add(lootTable(VOID_FISHING_FISH).setWeight(70).setQuality(-1))
 		));
-	}
+
+		consumer.accept(VOID_FISHING_LOCATION_AMETHYST_BIOMES, lootTable().withPool(pool()
+			.add(item(AMETHYST_BUDFISH).setWeight(16))
+			.add(item(AMETHYST_SHARD).setWeight(5).apply(count(range(2, 4))))
+			.add(item(QUARTZ).setWeight(5).apply(count(range(3, 6))))
+			.add(item(AMETHYST_BLOCK).setWeight(3).apply(count(range(2, 3))))
+			.add(item(BUDDING_AMETHYST).setWeight(1))
+		));
+
+		consumer.accept(VOID_FISHING_LOCATION_FIERY_HILLS, lootTable().withPool(pool()
+			.add(item(FLAREFIN_KOI).setWeight(20))
+			.add(item(BLAZE_ROD).setWeight(3).apply(count(range(1, 2))))
+			.add(item(MAGMA_BLOCK).setWeight(6).apply(count(range(2, 4))))
+			.add(item(GLOWSTONE_DUST).setWeight(7).apply(count(range(3, 7))))
+			.add(item(GHAST_TEAR))
+		));
+
+		consumer.accept(VOID_FISHING_LOCATION_FLESH_TUNDRA, lootTable().withPool(pool()
+			.add(item(CRIMSON_TIGERFISH).setWeight(27))
+			.add(item(FLESHY_PIRANHA).setWeight(19))
+			.add(item(ROTTEN_FLESH).setWeight(16).apply(count(range(3, 6))))
+			.add(item(NETHER_WART_BLOCK).setWeight(12).apply(count(range(2, 3))))
+			.add(item(GUNPOWDER).setWeight(12).apply(count(range(1, 5))))
+			.add(item(BONE).setWeight(14).apply(count(range(1, 4))))
+		));
+
+		consumer.accept(VOID_FISHING_LOCATION_FROZEN_SPIKES, lootTable().withPool(pool()
+			.add(item(FROST_MINNOW).setWeight(7))
+			.add(item(ICE).apply(count(range(2, 4))))
+			.add(item(PACKED_ICE).apply(count(range(2, 4))))
+			.add(item(BLUE_ICE).apply(count(range(2, 3))))
+			.add(item(SNOW_BLOCK).apply(count(range(2, 4))))
+			.add(item(SNOWBALL).apply(count(range(3, 8))))
+		));
+
+		consumer.accept(VOID_FISHING_LOCATION_VANILLA_BIOMES, lootTable().withPool(pool()
+			.add(item(OVERGROWN_COD).setWeight(8))
+			.add(item(CHORUS_FLOWER).setWeight(3).apply(count(range(1, 2))))
+			.add(item(ENDER_EYE).setWeight(2).apply(count(range(1, 2))))
+			.add(item(CHORUS_PLANT).setWeight(4).apply(count(range(3, 8))))
+		));
+
+		consumer.accept(VOID_FISHING_LOCATION_PRISMARINE_FOREST, lootTable().withPool(pool()
+			.add(item(BUBBLEFISH).setWeight(12))
+			.add(item(PRISMARINE).setWeight(6).apply(count(range(3, 5))))
+			.add(item(PRISMARINE_CRYSTALS).setWeight(5).apply(count(range(2, 6))))
+			.add(item(GLOW_INK_SAC).setWeight(5).apply(count(range(1, 2))))
+		));
+
+		consumer.accept(VOID_FISHING_LOCATION_THE_HALLOW, lootTable().withPool(pool()
+			.add(item(PRISMITE).setWeight(25))
+			.add(item(STARLIGHT_SOOT).setWeight(9).apply(count(range(2, 4))))
+			.add(item(CHERRY_LEAVES).setWeight(6).apply(count(range(2, 5))))
+			.add(item(PINK_PETALS).setWeight(6).apply(count(range(3, 6))))
+		));
+
+		consumer.accept(VOID_FISHING_LOCATION_WARPED_MARSH, lootTable().withPool(pool()
+			.add(item(GOOSH).setWeight(20))
+			.add(item(MOSS_BLOCK).setWeight(3).apply(count(range(3, 5))))
+			.add(item(VERDANT_FROGLIGHT).setWeight(6).apply(count(range(1, 2))))
+			// extra
+			.add(item(ASHEN_FROGLIGHT).setWeight(6).apply(count(range(1, 2))))
+		));
 
 
-	@Override
-	public void generate(@NonNull BiConsumer<ResourceKey<LootTable>, LootTable.Builder> consumer) {
-		define(registryLookup.join());
-		for (var entry : LOOT_TABLES.entrySet()) {
-			consumer.accept(entry.getKey(), entry.getValue());
-		}
 	}