+++ src/main/java/prismatic/shards/stellarity/key/StellarityConfiguredFeatures.java
@@ -26,6 +26,8 @@ public interface StellarityConfiguredFeatures {
 	ResourceKey<ConfiguredFeature<?, ?>> END_HIGHLANDS_SMALL_DIRT_PATCH = id("end_highlands/small_dirt_patch");
 	ResourceKey<ConfiguredFeature<?, ?>> END_HIGHLANDS_CHORUS_BUD = id("end_highlands/chorus_bud");
 	ResourceKey<ConfiguredFeature<?, ?>> END_HIGHLANDS_PITCHER_PLANT = id("end_highlands/pitcher_plant");
+	ResourceKey<ConfiguredFeature<?, ?>> END_HIGHLANDS_GRASS = id("end_highlands/grass");
+	ResourceKey<ConfiguredFeature<?, ?>> END_HIGHLANDS_ROOTS = id("end_highlands/roots");
 
 	ResourceKey<ConfiguredFeature<?, ?>> AMETHYST_FOREST_CALCITE_BOTTOM = id("amethyst_forest/calcite_bottom");
 	ResourceKey<ConfiguredFeature<?, ?>> AMETHYST_FOREST_AMETHYST_GEODE = id("amethyst_forest/amethyst_geode");