+++ src/main/java/prismatic/shards/stellarity/key/StellarityLootTables.java
@@ -13,6 +13,14 @@ public interface StellarityLootTables {
 	ResourceKey<LootTable> VOID_FISHING_TREASURE = id("void_fishing/treasure");
 	ResourceKey<LootTable> VOID_FISHING_FISHER_OF_VOIDS = id("void_fishing/fisher_of_voids");
 	ResourceKey<LootTable> VOID_FISHING_EVENT = id("void_fishing/event");
+	ResourceKey<LootTable> VOID_FISHING_LOCATION_AMETHYST_BIOMES = id("void_fishing/location/amethyst_biomes");
+	ResourceKey<LootTable> VOID_FISHING_LOCATION_FIERY_HILLS = id("void_fishing/location/fiery_hills");
+	ResourceKey<LootTable> VOID_FISHING_LOCATION_FLESH_TUNDRA = id("void_fishing/location/flesh_tundra");
+	ResourceKey<LootTable> VOID_FISHING_LOCATION_FROZEN_SPIKES = id("void_fishing/location/frozen_spikes");
+	ResourceKey<LootTable> VOID_FISHING_LOCATION_VANILLA_BIOMES = id("void_fishing/location/vanilla_biomes");
+	ResourceKey<LootTable> VOID_FISHING_LOCATION_PRISMARINE_FOREST = id("void_fishing/location/prismarine_forest");
+	ResourceKey<LootTable> VOID_FISHING_LOCATION_THE_HALLOW = id("void_fishing/location/the_hallow");
+	ResourceKey<LootTable> VOID_FISHING_LOCATION_WARPED_MARSH = id("void_fishing/location/warped_marsh");
 
 	static ResourceKey<LootTable> id(String id) {
 		return Stellarity.key(Registries.LOOT_TABLE, id);