+++ src/main/java/prismatic/shards/stellarity/key/StellarityPlacedFeatures.java
@@ -31,6 +31,8 @@ public interface StellarityPlacedFeatures {
 	ResourceKey<PlacedFeature> END_HIGHLANDS_CHORUS_BUDS = id("end_highlands/chorus_buds");
 	ResourceKey<PlacedFeature> END_HIGHLANDS_CHORUS_PLANTS = id("end_highlands/chorus_plants");
 	ResourceKey<PlacedFeature> END_HIGHLANDS_PITCHER_PLANTS = id("end_highlands/pitcher_plants");
+	ResourceKey<PlacedFeature> END_HIGHLANDS_GRASS = id("end_highlands/grass");
+	ResourceKey<PlacedFeature> END_HIGHLANDS_ROOTS = id("end_highlands/roots");
 
 	ResourceKey<PlacedFeature> AMETHYST_FOREST_CALCITE_BOTTOM = id("amethyst_forest/calcite_bottom");
 	ResourceKey<PlacedFeature> AMETHYST_FOREST_AMETHYST_GEODES = id("amethyst_forest/amethyst_geodes");